third_party/WebKit/LayoutTests/platform/win-xp/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left is 8
PASS rects[0].top is 8
PASS rects[0].width is 400
PASS rects[0].height is 160
PASS rects[1].left is 8
FAIL rects[1].top should be 19. Was 18.
FAIL rects[1].width should be 396. Was 378.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 59. Was 58.
FAIL rects[2].width should be 398. Was 378.
FAIL rects[2].height should be 18. Was 19.
PASS rects[3].left is 8
FAIL rects[3].top should be 99. Was 98.
FAIL rects[3].width should be 360. Was 380.
FAIL rects[3].height should be 18. Was 19.
PASS rects[4].left is 8
FAIL rects[4].top should be 139. Was 138.
FAIL rects[4].width should be 306. Was 247.
FAIL rects[4].height should be 18. Was 19.
Test 2
PASS rects.length is 4
PASS rects[0].left is 8
FAIL rects[0].top should be 215. Was 218.
FAIL rects[0].width should be 396. Was 378.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 255. Was 258.
FAIL rects[1].width should be 398. Was 378.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 295. Was 298.
FAIL rects[2].width should be 360. Was 380.
FAIL rects[2].height should be 18. Was 19.
PASS rects[3].left is 8
FAIL rects[3].top should be 335. Was 338.
FAIL rects[3].width should be 306. Was 247.
FAIL rects[3].height should be 18. Was 19.
Test 3
PASS rects.length is 4
FAIL rects[0].left should be 80. Was 76.
FAIL rects[0].top should be 411. Was 418.
FAIL rects[0].width should be 324. Was 310.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 451. Was 458.
FAIL rects[1].width should be 398. Was 378.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 491. Was 498.
FAIL rects[2].width should be 360. Was 380.
FAIL rects[2].height should be 18. Was 19.
PASS rects[3].left is 8
FAIL rects[3].top should be 531. Was 538.
FAIL rects[3].width should be 88. Was 38.
FAIL rects[3].height should be 18. Was 19.
Test 4
FAIL rects.length should be 7. Was 6.
PASS rects[0].left is 8
FAIL rects[0].top should be 596. Was 608.
PASS rects[0].width is 400
FAIL rects[0].height should be 278. Was 238.
PASS rects[1].left is 8
FAIL rects[1].top should be 607. Was 618.
FAIL rects[1].width should be 396. Was 378.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 725. Was 736.
FAIL rects[2].width should be 242. Was 230.
FAIL rects[2].height should be 18. Was 19.
FAIL rects[3].left should be 326. Was 314.
FAIL rects[3].top should be 725. Was 736.
FAIL rects[3].width should be 44. Was 86.
FAIL rects[3].height should be 18. Was 19.
PASS rects[4].left is 8
FAIL rects[4].top should be 765. Was 776.
FAIL rects[4].width should be 341. Was 398.
FAIL rects[4].height should be 18. Was 19.
PASS rects[5].left is 8
FAIL rects[5].top should be 805. Was 816.
FAIL rects[5].width should be 366. Was 291.
FAIL rects[5].height should be 18. Was 19.
FAIL rects[6].left should be 8. Threw exception TypeError: Cannot read property 'left' of undefined
FAIL rects[6].top should be 845. Threw exception TypeError: Cannot read property 'top' of undefined
FAIL rects[6].width should be 67. Threw exception TypeError: Cannot read property 'width' of undefined
FAIL rects[6].height should be 18. Threw exception TypeError: Cannot read property 'height' of undefined
Test 5
PASS rects.length is 5
FAIL rects[0].left should be 80. Was 76.
FAIL rects[0].top should be 921. Was 896.
FAIL rects[0].width should be 324. Was 310.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 1039. Was 1014.
FAIL rects[1].width should be 242. Was 230.
FAIL rects[1].height should be 18. Was 19.
FAIL rects[2].left should be 250. Was 238.
FAIL rects[2].top should be 950. Was 926.
PASS rects[2].width is 76
PASS rects[2].height is 103
FAIL rects[3].left should be 326. Was 314.
FAIL rects[3].top should be 1039. Was 1014.
FAIL rects[3].width should be 44. Was 86.
FAIL rects[3].height should be 18. Was 19.
PASS rects[4].left is 8
FAIL rects[4].top should be 1079. Was 1054.
FAIL rects[4].width should be 284. Was 220.
FAIL rects[4].height should be 18. Was 19.
Test 6
PASS rects.length is 7
PASS rects[0].left is 8
FAIL rects[0].top should be 1224. Was 1164.
PASS rects[0].width is 400
PASS rects[0].height is 200
PASS rects[1].left is 8
FAIL rects[1].top should be 1235. Was 1174.
FAIL rects[1].width should be 396. Was 378.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 1275. Was 1214.
FAIL rects[2].width should be 57. Was 54.
FAIL rects[2].height should be 18. Was 19.
FAIL rects[3].left should be 105. Was 102.
FAIL rects[3].top should be 1275. Was 1214.
FAIL rects[3].width should be 44. Was 43.
FAIL rects[3].height should be 18. Was 19.
PASS rects[4].left is 8
FAIL rects[4].top should be 1315. Was 1254.
FAIL rects[4].width should be 387. Was 363.
FAIL rects[4].height should be 18. Was 19.
PASS rects[5].left is 8
FAIL rects[5].top should be 1355. Was 1294.
FAIL rects[5].width should be 397. Was 389.
FAIL rects[5].height should be 18. Was 19.
PASS rects[6].left is 8
FAIL rects[6].top should be 1395. Was 1334.
FAIL rects[6].width should be 171. Was 148.
FAIL rects[6].height should be 18. Was 19.
Test 7
PASS rects.length is 7
PASS rects[0].left is 8
FAIL rects[0].top should be 1471. Was 1414.
FAIL rects[0].width should be 396. Was 378.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 1500. Was 1444.
PASS rects[1].width is 400
PASS rects[1].height is 40
PASS rects[2].left is 8
FAIL rects[2].top should be 1511. Was 1454.
FAIL rects[2].width should be 57. Was 54.
FAIL rects[2].height should be 18. Was 19.
FAIL rects[3].left should be 105. Was 102.
FAIL rects[3].top should be 1511. Was 1454.
FAIL rects[3].width should be 44. Was 43.
FAIL rects[3].height should be 18. Was 19.
PASS rects[4].left is 8
FAIL rects[4].top should be 1551. Was 1494.
FAIL rects[4].width should be 387. Was 363.
FAIL rects[4].height should be 18. Was 19.
PASS rects[5].left is 8
FAIL rects[5].top should be 1591. Was 1534.
FAIL rects[5].width should be 397. Was 389.
FAIL rects[5].height should be 18. Was 19.
PASS rects[6].left is 8
FAIL rects[6].top should be 1631. Was 1574.
FAIL rects[6].width should be 171. Was 148.
FAIL rects[6].height should be 18. Was 19.
Test 8
PASS rects.length is 4
FAIL rects[0].left should be 168. Was 168.5107421875.
FAIL rects[0].top should be 1678. Was 1625.7960205078125.
FAIL rects[0].width should be 269. Was 257.528564453125.
FAIL rects[0].height should be 316. Was 301.7777099609375.
FAIL rects[1].left should be 137. Was 137.8689727783203.
FAIL rects[1].top should be 1704. Was 1651.5074462890625.
FAIL rects[1].width should be 271. Was 257.528564453125.
FAIL rects[1].height should be 317. Was 301.77783203125.
FAIL rects[2].left should be 107. Was 107.22718811035156.
FAIL rects[2].top should be 1729. Was 1677.218994140625.
FAIL rects[2].width should be 246. Was 258.81414794921875.
FAIL rects[2].height should be 289. Was 303.309814453125.
FAIL rects[3].left should be 76. Was 76.58541107177734.
FAIL rects[3].top should be 1755. Was 1702.9305419921875.
FAIL rects[3].width should be 212. Was 173.32339477539062.
FAIL rects[3].height should be 247. Was 201.4259033203125.
Test 9
FAIL rects.length should be 1. Was 0.
FAIL rects[0].left should be 8. Threw exception TypeError: Cannot read property 'left' of undefined
FAIL rects[0].top should be 1903. Threw exception TypeError: Cannot read property 'top' of undefined
FAIL rects[0].width should be 0. Threw exception TypeError: Cannot read property 'width' of undefined
FAIL rects[0].height should be 18. Threw exception TypeError: Cannot read property 'height' of undefined
Test 9b
PASS rects.length is 1
PASS rects[0].left is 8
FAIL rects[0].top should be 1903. Was 1854.
PASS rects[0].width is 0
FAIL rects[0].height should be 18. Was 19.
Test 10
PASS rects.length is 4
FAIL rects[0].left should be 19. Was 18.
FAIL rects[0].top should be 2088. Was 2044.
FAIL rects[0].width should be 18. Was 19.
FAIL rects[0].height should be 393. Was 399.
FAIL rects[1].left should be 59. Was 58.
FAIL rects[1].top should be 2088. Was 2044.
FAIL rects[1].width should be 18. Was 19.
FAIL rects[1].height should be 377. Was 398.
FAIL rects[2].left should be 99. Was 98.
FAIL rects[2].top should be 2088. Was 2044.
FAIL rects[2].width should be 18. Was 19.
FAIL rects[2].height should be 372. Was 380.
FAIL rects[3].left should be 139. Was 138.
FAIL rects[3].top should be 2088. Was 2044.
FAIL rects[3].width should be 18. Was 19.
FAIL rects[3].height should be 360. Was 247.
Test 11
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
